restructure frontend game folder
